// apps/aggiemap/src/app/types/types.ts
export interface TemplateRendererOptions {
  /**
   * Text template. Anything between and including double matching curly braces will be replaced
   * with either a static value or a value based on object property lookup.
   *
   * @type {string}
   * @memberof TemplateRendererOptions
   */
  template?: string;

  /**
   * Object used for property lookup. Matched values will be used for replacement instead of a
   * static provided replacement value.
   *
   * @type {object}
   * @memberof TemplateRendererOptions
   */
  lookup?: object;

  /**
   * Replacement value.
   *
   * Only required in when providing an input string.
   *
   * If using inputObject, values will be derived from property value lookup.
   *
   *
   * @type {string}
   * @memberof TemplateRendererOptions
   */
  replacement?: string;
}

/**
 * Basic representation of a geographic point with longitude (x) and latitude(y) values
 *
 * @export
 * @interface Point
 */
export interface Point {
  latitude: number;
  longitude: number;
}